Program to count digits with catagorisation
The program takes input in a string, and traverses to find numbers from 0-9. It stores the frequency of each no. in it's respective position.
That is,
number of occurance of i is stored in array[i]
After full traversal, an array output is given showing the frequency of each number of the range.